Ignore config.c #chonkerkeys/10
